Create LICENSE
You only specifiy the license of the book's content (CC-BY-SA 4.0) in the [preface](https://bookdown.org/yihui/rmarkdown/) of the book. This first confused me a lot: After submitting a PR and transfering copyright of it to the book's authors I was wondering if the book was actually open source.

Besides allowing visitors of the repository to immediately determine the license of the project (without having to look at the book's preface), specifying the license in a separate file has the advantage of being machine-readable and thus being auto-detected by GitHub (as well as external entities).

I copied the license text 1:1 from [here](https://creativecommons.org/licenses/by-nc-sa/4.0/legalcode.txt).